add damage component

+++ Assets/Scripts/Components/HellthComponent.cs
@@ -5,13 +5,13 @@ using Entitas.CodeGeneration.Attributes;
 [Core, Event(EventTarget.Self)]
 public class Hellth : IComponent
 {
-    public float curValue;
-    public float maxValue;
+    public int curValue;
+    public int maxValue;
 }
 
 public partial class CoreEntity
 {
-    public void ReplaceHellth(float newCurValue)
+    public void ReplaceHellth(int newCurValue)
     {
         ReplaceHellth(newCurValue, hellth?.maxValue ?? newCurValue);
     }
